fix #21 Implemented Delete of CRUD

+++ src/main/java/com/example/enfermeria/dao/NurseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.enfermeria.entity.Nurse;
 
-public interface NurseRepository extends CrudRepository<Nurse, Integer> {
+public interface NurseRepository extends CrudRepository<Nurse, Long> {
 
 	Optional<Nurse> findByNameIgnoreCase(String name);
